test(ci): guard react package exports too via shared scripts/check-exports.mjs
