+++ .github/workflows/react.yaml
@@ -38,6 +38,9 @@ jobs:
       - name: Build the core
         run: npm run --workspace @simplepdf/embed build
 
+      - name: Build react-embed-pdf
+        run: npm run --workspace @simplepdf/react-embed-pdf build
+
       - name: Formatting
         run: npm run --workspace @simplepdf/react-embed-pdf test:format
 
@@ -46,3 +49,6 @@ jobs:
 
       - name: Tests
         run: npm run --workspace @simplepdf/react-embed-pdf test
+
+      - name: Export load
+        run: npm run --workspace @simplepdf/react-embed-pdf check:exports
