+++ embed/package.json
@@ -63,7 +63,8 @@
     "pretest": "npm run generate",
     "test": "vitest run",
     "test:watch": "vitest",
-    "check:size": "npm run build && node scripts/check-bundle-size.mjs"
+    "check:size": "npm run build && node scripts/check-bundle-size.mjs",
+    "check:exports": "node ../scripts/check-exports.mjs ."
   },
   "peerDependencies": {
     "@tanstack/ai": "^0.38.0",
